dtype= float

+++ plot_result.py
@@ -23,9 +23,9 @@ import numpy as np
 def plot_file(filename, data):
 	#f = open("drag_ligt.m", "r")
 	#s = f.read().split('\n')
-	time = np.array(data[0])
-	lift = np.array(data[1])
-	drag = np.array(data[2])
+	time = np.array(data[0], dtype= float)
+	lift = np.array(data[1], dtype= float)
+	drag = np.array(data[2], dtype= float)
 	#for line in s[1:]:
 	#	l = line.split(' ')
 	#	elemlist = []
